refactor login page
