+++ apps/web/src/components/LoginPageEditor/index.ts
@@ -1,0 +1,1 @@
+export * from './LoginPageEditor';
