+++ apps/web/src/components/LoginPageEditor/sections/LeftPanelThemeCard.tsx
@@ -1,0 +1,74 @@
+import { Card } from '@douglasneuroinformatics/libui/components';
+import { useTranslation } from '@douglasneuroinformatics/libui/hooks';
+import { LOGIN_THEMES } from '@opendatacapture/schemas/setup';
+import type { BrandingConfig } from '@opendatacapture/schemas/setup';
+
+import { getLoginGradient } from '@/utils/branding';
+
+import { THEME_LABELS } from '../constants';
+import { ColorField } from '../fields/ColorField';
+import { SwatchButton } from '../fields/SwatchButton';
+
+import type { BrandingEditor } from '../useBrandingForm';
+
+/** Left-panel background gradient: a theme swatch grid plus custom hex inputs. */
+export const LeftPanelThemeCard = ({ editor }: { editor: BrandingEditor }) => {
+  const { t } = useTranslation();
+  const { form, update } = editor;
+  return (
+    <Card>
+      <Card.Header>
+        <Card.Title>
+          {t({ en: 'Background Theme - Left Panel', fr: "Thème d'arrière-plan - Panneau gauche" })}
+        </Card.Title>
+        <Card.Description>
+          {t({
+            en: 'The gradient shown behind the branding panel.',
+            fr: 'Le dégradé affiché derrière le panneau.'
+          })}
+        </Card.Description>
+      </Card.Header>
+      <Card.Content className="flex flex-col gap-4">
+        <div className="grid grid-cols-2 gap-3 sm:grid-cols-4">
+          {LOGIN_THEMES.map((theme) => {
+            const swatchBranding: BrandingConfig =
+              theme === 'custom'
+                ? {
+                    customPrimaryColor: form.customPrimaryColor,
+                    customSecondaryColor: form.customSecondaryColor,
+                    loginTheme: 'custom'
+                  }
+                : { loginTheme: theme };
+            return (
+              <SwatchButton
+                isSelected={form.loginTheme === theme}
+                key={theme}
+                label={t(THEME_LABELS[theme])}
+                style={{ backgroundImage: getLoginGradient(swatchBranding) }}
+                onClick={() => update('loginTheme', theme)}
+              />
+            );
+          })}
+        </div>
+        {form.loginTheme === 'custom' && (
+          <div className="grid gap-4 sm:grid-cols-2">
+            <ColorField
+              id="customPrimaryColor"
+              label={t({ en: 'Gradient Start', fr: 'Début du dégradé' })}
+              swatchFallback="#000000"
+              value={form.customPrimaryColor}
+              onChange={(v) => update('customPrimaryColor', v)}
+            />
+            <ColorField
+              id="customSecondaryColor"
+              label={t({ en: 'Gradient End', fr: 'Fin du dégradé' })}
+              swatchFallback="#000000"
+              value={form.customSecondaryColor}
+              onChange={(v) => update('customSecondaryColor', v)}
+            />
+          </div>
+        )}
+      </Card.Content>
+    </Card>
+  );
+};
